refactoring: move shouldFeed fallback logic to usecase, change eventbus logic(one sub), move event_handler to usecase
