--- modules/040-node-manager/images/fencing-agent/src/internal/adapters/kubeclient/fallback.go
@@ -1,18 +1,0 @@
-package kubeclient
-
-import (
-	"context"
-
-	v1meta "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	"fencing-agent/internal/lib/logger/sl"
-)
-
-func (c *Client) ShouldFeed(ctx context.Context) bool {
-	_, err := c.client.CoreV1().Nodes().List(ctx, v1meta.ListOptions{})
-	if err != nil {
-		c.logger.Debug("kubernetes API is not available", sl.Err(err))
-		return false
-	}
-	return true
-}
