+++ modules/040-node-manager/images/fencing-agent/src/internal/domain/node_group_state.go
@@ -1,0 +1,6 @@
+package domain
+
+type NodeGroupState struct {
+	TotalNodes int   `json:"total_nodes"`
+	Timestamp  int64 `json:"timestamp"`
+}
