+++ modules/040-node-manager/images/fencing-agent/src/internal/usecase/eventbus.go
@@ -2,73 +2,67 @@ package usecase
 
 import (
 	"context"
-	"sync"
-	"sync/atomic"
-
 	"fencing-agent/internal/domain"
+	"fmt"
+	"sync"
 )
 
 type EventsBus struct {
-	subscribers map[chan domain.Event]*subscriber
-	mu          sync.RWMutex
+	events    chan domain.Event
+	ciliumSub *cilium
+	mu        sync.Mutex
 }
 
-type subscriber struct {
-	ch     chan domain.Event
-	ctx    context.Context
-	closed atomic.Bool
+type cilium struct {
+	ctx context.Context
 }
 
 func NewEventsBus() *EventsBus {
 	return &EventsBus{
-		subscribers: make(map[chan domain.Event]*subscriber),
+		events:    make(chan domain.Event),
+		ciliumSub: nil,
 	}
 }
 
 func (e *EventsBus) Publish(event domain.Event) {
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-
-	for _, sb := range e.subscribers {
-		if sb.closed.Load() {
-			continue
-		}
-		select {
-		case <-sb.ctx.Done():
-		case sb.ch <- event:
-		}
-	}
-}
-
-func (e *EventsBus) Subscribe(ctx context.Context) <-chan domain.Event {
 	e.mu.Lock()
 	defer e.mu.Unlock()
-
-	ch := make(chan domain.Event)
-	sub := &subscriber{
-		ch:  ch,
-		ctx: ctx,
-	}
-	e.subscribers[ch] = sub
-
-	go func() {
-		<-ctx.Done()
-		e.unsubscribe(ch)
-	}()
-	return ch
-}
-
-func (e *EventsBus) unsubscribe(ch chan domain.Event) {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-
-	sub, exists := e.subscribers[ch]
-	if !exists {
+	if e.ciliumSub == nil {
 		return
 	}
 
-	if sub.closed.CompareAndSwap(false, true) {
-		delete(e.subscribers, ch)
-		close(sub.ch)
+	select {
+	case e.events <- event:
+	case <-e.ciliumSub.ctx.Done():
+		return
 	}
+}
+
+func (e *EventsBus) Subscribe(ctx context.Context) (<-chan domain.Event, error) {
+	e.mu.Lock()
+	defer e.mu.Unlock()
+
+	if e.ciliumSub != nil {
+		return nil, fmt.Errorf("already subscribed")
+	}
+
+	e.ciliumSub = &cilium{ctx: ctx}
+
+	go func() {
+		<-ctx.Done()
+		e.unsubscribe()
+	}()
+
+	return e.events, nil
+}
+
+func (e *EventsBus) unsubscribe() {
+	e.mu.Lock()
+	defer e.mu.Unlock()
+
+	if e.ciliumSub == nil {
+		return
+	}
+
+	e.ciliumSub = nil
 }
